# docs/OpenApi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8081
    description: Generated server url
...
paths:
...
  /api/v1/categories/{categoryId}/items/{itemId}:
    get:
      tags:
        - Item Management
      summary: Retrieve a specific item
      description: Retrieves the item with the specified ID within the specified category.
      operationId: getItem
      parameters:
        - name: categoryId
          in: path
          description: ID of the category
          required: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: path
          description: ID of the item to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content: { }
    put:
      tags:
        - Item Management
      summary: Update an existing item
      description: >-
        Updates the item with the specified ID within the specified category
        using the provided item data.
      operationId: updateItem
      parameters:
        - name: categoryId
          in: path
          description: ID of the category
          required: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: path
          description: ID of the item to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDto'
        required: true
      responses:
        '200':
          description: OK
          content: { }
    delete:
      tags:
        - Item Management
      summary: Delete a specific item
      description: Deletes the item with the specified ID within the specified category.
      operationId: deleteItem
      parameters:
        - name: categoryId
          in: path
          description: ID of the category
          required: true
          schema:
            type: integer
            format: int64
        - name: itemId
          in: path
          description: ID of the item to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content: { }
...
  /api/v1/categories/{categoryId}/items:
    get:
      tags:
        - Item Management
      summary: Retrieve all items
      description: >-
        Retrieves all items within the specified category that match the
        provided filters.
      operationId: getAllItems
      parameters:
        - name: categoryId
          in: path
          description: ID of the category
          required: true
          schema:
            type: integer
            format: int64
        - name: itemFilter
          in: query
          description: Filter criteria for items
          required: true
          schema:
            $ref: '#/components/schemas/ItemFilter'
      responses:
        '200':
          description: OK
          content: { }
    post:
      tags:
        - Item Management
      summary: Create a new item
      description: >-
        Creates a new item within the specified category using the provided item
        data.
      operationId: createItem
      parameters:
        - name: categoryId
          in: path
          description: ID of the category to which the item belongs
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDto'
        required: true
      responses:
        '200':
          description: OK
          content: { }
...
components:
  schemas:
...
    ItemDto:
      type: object
      properties:
        itemId:
          type: integer
          description: Unique identifier for the item.
          format: int64
          example: 1
        categoryId:
          type: integer
          description: Unique identifier for the category to which the item belongs.
          format: int64
          example: 10
        itemName:
          type: string
          description: Name of the item.
          example: Buy groceries
        itemStatus:
          type: string
          description: Current status of the item.
          example: In Progress
        itemNotes:
          type: string
          description: Additional notes or description for the item.
          example: Need to buy fruits and vegetables.
      description: Details of the item to be created
...
    ItemFilter:
      type: object
      properties:
        page:
          type: integer
          description: Page number for pagination.
          format: int32
          example: 0
          default: 0
        size:
          type: integer
          description: Number of items per page for pagination.
          format: int32
          example: 2
          default: 2
        categoryId:
          type: integer
          description: Unique identifier for the category to which the items belong.
          format: int64
          example: 10
        itemName:
          type: string
          description: Name of the item to filter by.
          example: Buy groceries
        itemStatus:
          type: string
          description: Current status of the item to filter by.
          example: In Progress
        direction:
          type: string
          description: Sorting direction for the item list.
          example: ASC
          default: ASC
          enum:
            - ASC
            - DESC
      description: >-
        Data Transfer Object for filtering items, including pagination and
        sorting.